clear some clippy warnings

+++ state-machines/binary/src/binary_add_collector.rs
@@ -64,7 +64,7 @@ impl BusDevice<u64> for BinaryAddCollector {
         _pending: &mut VecDeque<(BusId, Vec<u64>)>,
     ) -> bool {
         debug_assert!(*bus_id == OPERATION_BUS_ID);
-        let instance_complete = self.inputs.len() == self.num_operations as usize;
+        let instance_complete = self.inputs.len() == self.num_operations;
 
         if instance_complete && !self.force_execute_to_end {
             return false;
@@ -97,7 +97,7 @@ impl BusDevice<u64> for BinaryAddCollector {
 
         self.inputs.push([OperationBusData::get_a(&op_data), OperationBusData::get_b(&op_data)]);
 
-        self.inputs.len() < self.num_operations as usize || self.force_execute_to_end
+        self.inputs.len() < self.num_operations || self.force_execute_to_end
     }
 
     /// Returns the bus IDs associated with this instance.
